requirements/base.txt
# Base
Pillow==7.1.0
pycparser==2.20
WeasyPrint==51
psycopg2==2.8.3
phonenumbers==8.12.6

# Django
Django==2.2.14
djangorestframework==3.9.4
django-extras==0.3
django-phonenumber-field==2.3.0

# Enviroment
django-environ==0.4.5

# Password security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Celery
redis>=3.2.0
django-redis==4.10.0
tornado>=5.0.0,<6.0.0
celery==4.2.1
flower==0.9.2

# Git
culqipy==0.2.6
pre-commit==2.7.1
